Implemented functionality to Add to cart button
When the user clicks the Add to cart button it will be saved to the database and if he presses it again on the same product the quantity will be increased by +1.
In short, sending product data from API to database when a user presses add to cart button so we can retrieve the details in cart fragment to calculate the total bill.